replace tet4 test with solenoid test (remove duplicated work)
